Add failing test for AgentEmbedderMojo.adjustPathSeparatorToHostFS
If anyone ever asks me again, if it really makes sense to cover a
seemingly trivial one-line method with tests, run this test to find out
why it fails:

The method replaces path separators, so they fit the target file system.
But critically, the method contains an "optimisation" and only performs
the replacement if File.separator is unequal to the target FS separator,
assuming that the given path only contains path separators identical to
File.separator.

In practice, this will be the case, if(!) the string passed into this
method came from a File instance. But as nobody can guarantee that, the
input parameters could be such, that File.separator does not match the
path. E.g., we are on Windows with File.separator being a backslash, but
the passed-in path string contains forward slashes. In that case, there
would falsely be no conversion due to the ternary expression with the
premature optimisation in the code.

The next commit will fix the method, making the "unnecessary" test pass.