ZipFileSystemTool handles nested zips on JDK < 12
The new tool class is an application class, no longer a set of helper
methods in tests only. As on older JDKs nested zips are not supported,
class ZipFileSystemTool.VirtualNestedZipFileSystem now
transparently represents a zip FS based on a temp-file copied to the
default FS to provide nested zip support on JDKs 8-11.