Add JavaAgentLauncher for multi-agent support
The launcher adds a feature not supported by 'Launcher-Agent-Class' in
the JRE out of the box: It is possible to embed and start multiple
agents using the new class JavaAgentLauncher as a dispatcher. Equally
unsupported by the JRE is the option to specify an option string for
each agent.

These two features brings the agent embedder up to par with starting
agents from the JVM command line via '-javaagent', because there
multiple agents and an option string per agent are possible. Both might
be necessary in some cases.

To do: Add UTs for the launcher and an IT for multiple agents.